ディレクトリ構成の修正

+++ 1a_create_disk.sh
@@ -12,8 +12,10 @@ source "$_DIR/conf/conf_part.sh"
 source "$_DIR/com/com.sh"
 
 # ディスク切断
-unset_device "$_DISK_IMG_PATH"
-sudo rm "$_DISK_IMG_PATH"
+if [ -e "$_DISK_IMG_PATH" ]; then
+    unset_device "$_DISK_IMG_PATH"
+    sudo rm "$_DISK_IMG_PATH"
+fi
 
 # ディスク作成
 dd if=/dev/zero of="$_DISK_IMG_PATH" bs=1G count="$_TOTAL_GSIZE"
